Remove variables only used with .DEFERRED_INIT
In PR109087 it was noticed that we rely on DSE to remove .DEFERRED_INIT
when it is the only remaining use of a variable.  Since DSE is imperfect
and even if it were not would be limited by the amount of statements to
walk the following enhances the unused var removal pass to handle
.DEFERRED_INIT like CLOBBERs, thus we do not keep local variables just
because they are deferred initialized.

Bootstrapped and tested on x86_64-unknown-linux-gnu, for now queued
for stage1 unless somebody thinks its applicable for GCC 13.

	* tree-ssa-live.cc (remove_unused_locals): Do not treat
	the .DEFERRED_INIT of a variable as use, instead remove
	that if it is the only use.

	* gcc.dg/auto-init-unused-1.c: New testcase.