c++: redeclaring member of constrained class template [PR96830]
When redeclaring a member of a constrained class template, we need to
repeat the class's constraints, but it turns out we don't verify
anywhere that the repeated constraints match the class's constraints.
A safe place to do so seems to be in push_template_decl, nearby a
similar consistency check for the template parameter list length of a
redeclaration.

Bootstrapped and regtested on x86_64-pc-linux-gnu, does this look OK for
trunk/12?

	PR c++/96830

gcc/cp/ChangeLog:

	* pt.cc (push_inline_template_parms_recursive): Set
	TEMPLATE_PARMS_CONSTRAINTS.
	(push_template_decl): For a redeclaration, verify constraints
	for each enclosing template scope match those of the original
	template declaratation.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp2a/concepts-class5.C: New test.
	* g++.dg/cpp2a/concepts-class5a.C: New test.