fix array element count
This test is similar to pr103116-1.c, but instead of writing to
4*COUNT elements of x, it writes to 8*COUNT elements, but the
definition of x seems to have been adjusted along with the loop.  Fix
the array size so that it doesn't scribble over unrelated
statically-allocated objects.

Regstrapped on ppc64-linux-gnu.  Also tested with gcc-11 on vxworks7r2
(x86- and x86_64-), where the scribbling caused visible runtime effects.
Ok to install?  I'm tempted to put this in as obvious.

for  gcc/testsuite/ChangeLog

	* gcc.dg/vect/pr103116-2.c (x): Fix array size.