c++: generic lambda, local class, __func__ [PR108242]
Tested x86_64-pc-linux-gnu, applying to trunk.

-- 8< --

Here we are trying to do name lookup in a deferred instantiation of t() and
failing to find __func__.  tsubst_expr already tries to instantiate members
of local classes, but was failing with the partial instantiation of generic
lambdas.

	PR c++/108242

gcc/cp/ChangeLog:

	* pt.cc (tsubst_expr) [TAG_DEFN]: Handle partial instantiation.

gcc/testsuite/ChangeLog:

	* g++.dg/cpp1y/lambda-generic-func2.C: New test.